third_party/blink/web_tests/external/wpt/sanitizer-api/sanitizer-javascript-url-expected.txt

This is a testharness.js-based test.
[FAIL] setHTML testcase built-in-animating-url-attributes-list/2, "<svg><animateMotion attributeName="href"></svg>"
  assert_true: [object SVGSVGElement].isEqual([object SVGSVGElement]) fails. Path: <DIV> / <svg>. expected true got false
[FAIL] parseHTML testcase built-in-animating-url-attributes-list/2, "<svg><animateMotion attributeName="href"></svg>"
  assert_true: [object SVGSVGElement].isEqual([object SVGSVGElement]) fails. Path: [object HTMLDocument] / <HTML> / <BODY> / <svg>. expected true got false
[FAIL] setHTML testcase built-in-animating-url-attributes-list/3, "<svg><animateMotion attributeName="xlink:href"></svg>"
  assert_true: [object SVGSVGElement].isEqual([object SVGSVGElement]) fails. Path: <DIV> / <svg>. expected true got false
[FAIL] parseHTML testcase built-in-animating-url-attributes-list/3, "<svg><animateMotion attributeName="xlink:href"></svg>"
  assert_true: [object SVGSVGElement].isEqual([object SVGSVGElement]) fails. Path: [object HTMLDocument] / <HTML> / <BODY> / <svg>. expected true got false
Harness: the test ran to completion.
